Container Hosts/Projects register in Consul + refactoring
Reorganized a lot of code to do with external connections and moved them out of the connections package
Saves project metadata in Consul KV under its windlass_worker@hostname for below reasons
Worker will check Consul KV for any projects associated with it on startup and spin up goroutine for health check
Same topic, implemented Ping for LXD container host for said health check
On creation of a project, it is also registered in Consul, KV and service
LXD Host repo keeps in-memory the client TLS certs/keys while also persisting them in Vault for on next startup
EstablishConnections checks connections on startup instead. Do we wanna change this to be in degraded state rather than outright not running?